Merge version `0.1.9`

+++ dev_scripts/clear_pyinstaller_dist.py
@@ -1,3 +1,7 @@
+"""
+清理开发过程中测试运行时的打包中间文件及结果文件
+"""
+
 import os
 from pathlib import Path
 from shutil import rmtree
@@ -9,10 +13,14 @@ spec_path_list = list(src_dir_path.glob("*.spec"))
 
 
 def clear():
-    rmtree(dist_path)
-    rmtree(build_path)
-    for spec_file in spec_path_list:
-        os.remove(spec_file)
+    if dist_path.exists():
+        rmtree(dist_path)
+    if build_path.exists():
+        rmtree(build_path)
+    if spec_path_list:
+        for spec_file in spec_path_list:
+            os.remove(spec_file)
+    print("All cleaned.")
 
 
 clear()
